split the IP limiter from the proxy one
