+++ server/app.js
@@ -11,7 +11,8 @@ var express = require('express'),
   cors = require('cors'),
   jsonp = require('./jsonp'),
   redirector = require('./redirector'),
-  limiter = require('./limiter'),
+  ipLimiter = require('./limiters/ip'),
+  proxyLimiter = require('./limiters/proxy'),
   router = require('./router');
 
 var app = express();
@@ -30,7 +31,8 @@ app.use(cors({
 app.use(redirector.middleware);
 app.use(jsonp);
 app.use(express['static'](__dirname + '/..'));
-app.use(limiter());
+app.use(ipLimiter());
+app.use(proxyLimiter());
 app.use('/', router);
